Move FakeEarthdataLoginServer to its own file
