+++ packages/common/index.js
@@ -6,3 +6,4 @@ exports.task = require('./task');
 exports.cliUtils = require('./cli-utils');
 exports.CollectionConfigStore = require('./collection-config-store');
 exports.testUtils = require('./test-utils');
+exports.FakeEarthdataLoginServer = require('./fake-earthdata-login-server');
